pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "glide"
version = "0.1.0"
description = "GPU-accelerated Lightweight Ice Dynamics Engine"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "h5py>=3.0.0",
]

[project.optional-dependencies]
cuda = ["cupy-cuda12x>=12.0.0"]
data = ["netCDF4>=1.5.0", "rasterio>=1.2.0"]
all = ["cupy-cuda12x>=12.0.0", "netCDF4>=1.5.0", "rasterio>=1.2.0"]

[project.urls]
Repository = "https://github.com/username/glide"

[tool.setuptools.packages.find]
include = ["glide*"]

[tool.setuptools.package-data]
glide = ["cuda/*.cu"]
